playing field grid

+++ include/Types/Cell.hpp
@@ -4,6 +4,6 @@
 
 struct Cell
 {
-    sf::Vector2f size;
-    sf::Vector2f pos;
+    sf::Vector2i size;
+    sf::Vector2i pos;
 };
